refactor(errors): reorganize filestructure
