--- src/errors/index.ts
@@ -1,2 +1,0 @@
-export { classify } from "./classify";
-export { AgentCoreCLIError, InputValidationError } from "./types";
