+++ src/errors/index.tsx
@@ -1,0 +1,2 @@
+export { classify } from "./classify";
+export { AgentCoreCLIError, InputValidationError } from "./errors";
